refactor: 💡 flatten layout folder & move default id to constant
