+++ packages/mui-layout/src/components/HeaderOffset/HeaderOffset.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import useTheme from "@material-ui/core/styles/useTheme"
-import { useLayoutCtx } from "../../core/Context"
+import { useLayoutCtx } from "../../contexts"
 import HeaderOffsetCompiler from "../../compilers/HeaderOffsetCompiler"
 import { createBreakpointStyles } from "../../utils"
 import { generateStyledProxyCreator } from '../Shared/StyledProxy';
