--- packages/mui-layout/src/core/Context/index.tsx
@@ -1,3 +1,0 @@
-export * from "./LayoutContext"
-export * from "./SidebarContext"
-export * from "./WindowContext"
