--- packages/mui-layout/src/core/index.ts
@@ -1,5 +1,0 @@
-export { default } from '../builders';
-export * from './Context';
-export { default as useSidebar } from './hooks/useSidebar';
-export { default as useSidebarCta } from './hooks/useSidebarCta';
-export { default as useInsetSidebar } from './hooks/useInsetSidebar';
