+++ packages/mui-layout/src/hooks/useBreakpointConfig.ts
@@ -1,6 +1,6 @@
-import { MapBreakpoint } from "../../types"
+import { MapBreakpoint } from "../types"
 import useScreen from "./useScreen"
-import { pickNearestBreakpoint } from "../../utils"
+import { pickNearestBreakpoint } from "../utils"
 
 export const useBreakpointConfig = <T>(configMap: MapBreakpoint<T>) => {
   const screen = useScreen()
